requirements.txt
groq>=0.5.0
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
